backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-environ==0.13.0
openai==1.109.1
python-docx==1.2.0
psycopg[binary]==3.3.3
pypdf==6.1.1
